requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
markdown-it-py==2.2.0
mdurl==0.1.2
multidict==6.0.4
openai==0.27.6
prompt-toolkit==3.0.38
Pygments==2.15.1
python-dotenv==1.0.0
requests==2.30.0
rich==13.3.5
smmap==5.0.0
tqdm==4.65.0
urllib3==2.0.2
wcwidth==0.2.6
yarl==1.9.2
pytest==7.3.1
tiktoken
